Move userTSDB to a dedicated file
